upgrade upstream api: lessdb

+++ websrv/frontend/s2.go
@@ -188,7 +188,7 @@ func (c S2) IndexAction() {
 	}
 
 	if len(out) > 10 {
-		store.CacheSetBytes([]byte(hid), out, 36000+rand.Intn(36000))
+		store.CacheSetBytes([]byte(hid), out, uint32(36000+rand.Intn(36000)))
 	}
 
 	// fmt.Println("rez size", len(out))
